MTL-1319 show errors when basecamp wont start
Since 0.9.x, basecamp has not shown an error when the service fails to start.  This is due to the layering of services we have in place:

- go application
- running in a container
- behind systemd
- the systemd service just called `podman start`
- `podman start` returns a `0` status code even if the container failed to start
- the exit code of the container is also dependent upon whether or not the container started up

To resolve this, moved `podman start` into the `basecamp-init.sh` script.  This required some slight logic adjustments to the code.  To address the last bullet, I do a purge of the existing container before starting one and then I check the error code of that within the metadata.

I also added additional error checking because it's possible to run into a situation where more than one basecamp image exists, and thus the script failed to complete properly.  This accounts for that by removing things and starting a fresh container every time.

Signed-off-by: Jacob Salmela <jacob.salmela@hpe.com>