CASMINST-5622 Restore `DOCKER_ARGS`
Several meta labels are missing from the built image that were provided
via `$DOCKER_ARGS`, this is preventing basecamp's systemd service from
starting properly. The `DOCKER_ARGS` argument  was removed in #26.
The removal was a mistake, and `DOCKER_ARGS`  was restored in the `Jenkinsfile`
via #29 however the `Makefile` was missed. This change restores the
missing change in the `Makefile`.